requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
djangorestframework==3.8.2
idna==2.7
kombu==4.2.1
psycopg2-binary==2.7.5
pytz==2018.4
redis==2.10.6
requests==2.19.1
urllib3==1.23
vine==1.1.4
